Revert^2 "Update JOSDK to 5.1.0"
This reverts commit 0197c7588251d65c4603eb5f58f0d2c038a757e4.

Reason for revert: Tested the change on private cluster and it was working fine. I want to test the change in SAP's DEV landscape.

+++ operator/src/main/java/com/google/gerrit/k8s/operator/api/model/network/NetworkMemberWithSsh.java
@@ -18,7 +18,7 @@ import com.google.gerrit.k8s.operator.api.model.shared.HttpSshServiceConfig;
 import java.util.Objects;
 
 public class NetworkMemberWithSsh extends NetworkMember {
-  private int sshPort = 29418;
+  private long sshPort = 29418L;
 
   public NetworkMemberWithSsh() {}
 
@@ -27,11 +27,11 @@ public class NetworkMemberWithSsh extends NetworkMember {
     this.sshPort = serviceConfig.getSshPort();
   }
 
-  public int getSshPort() {
+  public long getSshPort() {
     return sshPort;
   }
 
-  public void setSshPort(int sshPort) {
+  public void setSshPort(long sshPort) {
     this.sshPort = sshPort;
   }
